+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.54.3] - 2019-05-07
+### Added
+- CXV1-18158 - Added new functions to create and update transfer-lists.
+
 ## [8.54.2] - 2019-05-06
 ### Added
 - CXV1-18151 - Add unit tests for typing and read indicators.